Resolves #187, #193: Prevent repeated action executions
