+++ lib/vagrant-hostsupdater/Action/UpdateHosts.rb
@@ -1,21 +1,11 @@
-require_relative "../HostsUpdater"
 module VagrantPlugins
   module HostsUpdater
     module Action
-      class UpdateHosts
-        include HostsUpdater
+      class UpdateHosts < BaseAction
 
-
-        def initialize(app, env)
-          @app = app
-          @machine = env[:machine]
-          @ui = env[:ui]
-        end
-
-        def call(env)
+        def run(env)
           @ui.info "[vagrant-hostsupdater] Checking for host entries"
           addHostEntries()
-          @app.call(env)
         end
 
       end
