+++ lib/vagrant-hostsupdater/version.rb
@@ -1,5 +1,5 @@
 module VagrantPlugins
   module HostsUpdater
-    VERSION = '1.1.1'
+    VERSION = '1.2.1'
   end
 end
